Sound scrambling using Circle Map and Wavelets
I used "level" for the depth of the tree , "nr_numbers" represents the numbers used by the circle map , "frame_samples_length" should be chosen depending on the size of the signal we want to encrypt. 
NPCR_and_UACI-is a function that will be used in the other types of chaotic maps.